finished support for nigma variables
