+++ dev/js/utils/parser.js
@@ -3,7 +3,6 @@ const Uniform = require('./variables/uniform');
 const Specific = require('./variables/specific');
 const Categorical = require('./variables/categorical');
 const Variable = require('./variables/variable');
-const ExpressionEvaluator = require('./variables/expression-evaluator');
 
 var VariableParser = {
   _detectType(nigmaCode) {
@@ -94,33 +93,7 @@ var VariableParser = {
     }
 
     return output;
-  },
-
-  isEvaluable(expression, variables) {
-    console.log({
-      expression: expression,
-      variables: variables
-    });
-    var evaluableVariables = Variable.retrieveEvaluableVariables(variables);
-    var match = expression.match(/\$[A-Za-z]/g) || [];
-    console.log(match);
-    var compoundOfEvaluable = match.every((varName) => evaluableVariables[varName] != null);
-    if (compoundOfEvaluable) {
-      var output = ExpressionEvaluator.evaluate(expression, evaluableVariables);
-      console.log(output);
-    } else {
-      return {
-        error: true,
-        message: "lala"
-      }
-    }
-  },
-
-
-
-
-
-
+  }
 }
 
 module.exports = VariableParser;
